clean-ups

+++ ph/src/fp/hash2.rs
@@ -59,6 +59,7 @@ impl<GS: GroupSize> FPHash2Conf<GS> {
 }
 
 impl<GS: GroupSize, SS: SeedSize> FPHash2Conf<GS, SS> {
+    /// Returns configuration that uses seeds and groups of sizes given in bits.
     pub fn bps_bpg(bits_per_seed: SS, bits_per_group: GS) -> Self {
         bits_per_seed.validate().unwrap();
         bits_per_group.validate().unwrap();
@@ -239,7 +240,7 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
         if !self.use_multiple_threads || !keys.has_par_for_each_key() {
             return self.build_array(keys, level_size_segments, level_size_groups, group_seed);
         }
-        let mut result = vec![0u64; level_size_segments].into_boxed_slice();
+        let mut result = vec![0u64; level_size_segments as usize].into_boxed_slice();
         let result_atom = AtomicU64::from_mut_slice(&mut result);
         let mut collision: Box<[AtomicU64]> = (0..level_size_segments).map(|_| AtomicU64::default()).collect();
         let level_seed = self.level_nr();
@@ -252,8 +253,8 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
     }
 
     /// Update `best_array` and `best_seeds` copying groups that are better (have more ones in `array`) from `array` and `array_seed`.
-    fn update_best<GetGroupSeed>(&self, level_size_groups: u32, best_array: &mut [u64], best_seeds: &mut [SS::VecElement], array: &[u64], array_seed: GetGroupSeed)
-        where GetGroupSeed: Fn(u32) -> u16
+    fn update_best<GetGroupSeed>(&self, level_size_groups: u64, best_array: &mut [u64], best_seeds: &mut [SS::VecElement], array: &[u64], array_seed: GetGroupSeed)
+        where GetGroupSeed: Fn(u64) -> u16
     {
         for group_index in 0..level_size_groups {
             self.conf.bits_per_group.conditionally_copy_group(best_array, array, group_index as usize,
@@ -266,7 +267,7 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
         }
     }
 
-    fn updated_best(&self, level_size_groups: u32, best: Seeds<SS::VecElement>, array2: Box<[u64]>, seed2: u16) -> Seeds<SS::VecElement>
+    fn updated_best(&self, level_size_groups: u64, best: Seeds<SS::VecElement>, array2: Box<[u64]>, seed2: u16) -> Seeds<SS::VecElement>
     {
         match best {
             Seeds::None => Seeds::Single(array2, seed2),
@@ -282,7 +283,7 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
         }
     }
 
-    fn select_best_seeds(&self, level_size_groups: u32, s1: Seeds<SS::VecElement>, s2: Seeds<SS::VecElement>) -> Seeds<SS::VecElement>
+    fn select_best_seeds(&self, level_size_groups: u64, s1: Seeds<SS::VecElement>, s2: Seeds<SS::VecElement>) -> Seeds<SS::VecElement>
     {
         match (s1, s2) {
             (Seeds::PerGroup(mut array1, mut seeds1), Seeds::PerGroup(array2, seeds2)) => {
@@ -300,14 +301,14 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
     /// Build (by calling `build_for_group`) arrays for all group seeds sequentially and select best groups and seeds (which are returned).
     /// `build_for_group` can use multiple threads internally to build each array.
     #[inline(always)]
-    fn best_array<AB>(&self, build_for_group: AB, level_size_groups: usize) -> (Box<[u64]>, Box<[SS::VecElement]>)
+    fn best_array<AB>(&self, build_for_group: AB, level_size_groups: u64) -> (Box<[u64]>, Box<[SS::VecElement]>)
         where AB: Fn(u16) -> Box<[u64]> // build array for given group nr
     {
         let mut best_array = build_for_group(0);
-        let mut best_seeds = self.conf.bits_per_seed.new_zeroed_seed_vec(level_size_groups);
+        let mut best_seeds = self.conf.bits_per_seed.new_zeroed_seed_vec(level_size_groups as usize);
         for group_seed in 1..=self.last_seed() {
             let with_new_seed = build_for_group(group_seed);
-            self.update_best(level_size_groups as u32, &mut best_array, &mut best_seeds, &with_new_seed, |_| group_seed);
+            self.update_best(level_size_groups, &mut best_array, &mut best_seeds, &with_new_seed, |_| group_seed);
         }
         (best_array, best_seeds)
     }
@@ -315,23 +316,23 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
     /// Build (by calling `build_for_group`) arrays for all group seeds and select best groups and seeds (which are returned).
     /// Multiple levels can be built in parallel and `build_for_group` should not use multiple threads internally.
     #[inline(always)]
-    fn best_array_mt<AB>(&self, build_for_group: AB, level_size_groups: usize) -> (Box<[u64]>, Box<[SS::VecElement]>)
+    fn best_array_mt<AB>(&self, build_for_group: AB, level_size_groups: u64) -> (Box<[u64]>, Box<[SS::VecElement]>)
         where AB: Fn(u16) -> Box<[u64]> + Sync
     {
         let s = (0..=self.last_seed()).into_par_iter().fold(|| Seeds::None::<SS::VecElement>, |best, seed| {
             let new_arr = build_for_group(seed);
-            self.updated_best(level_size_groups as u32, best, new_arr, seed)
+            self.updated_best(level_size_groups, best, new_arr, seed)
         }).reduce_with(|best, new| {
-            self.select_best_seeds(level_size_groups as u32, best, new)
+            self.select_best_seeds(level_size_groups, best, new)
         }).unwrap();
         match s {
-            Seeds::Single(a, seed) => (a, self.conf.bits_per_seed.new_seed_vec(seed, level_size_groups)),
+            Seeds::Single(a, seed) => (a, self.conf.bits_per_seed.new_seed_vec(seed, level_size_groups as usize)),
             Seeds::PerGroup(a, seeds) => (a, seeds),
             Seeds::None => unreachable!()
         }
     }
 
-    fn build_next_level_prehash<KS, K>(&mut self, keys: &mut KS, level_size_groups: usize, level_size_segments: usize)
+    fn build_next_level_prehash<KS, K>(&mut self, keys: &mut KS, level_size_groups: u64, level_size_segments: usize)
         where K: Hash + Sync, KS: KeySet<K> + Sync
     {
         let level_seed = self.level_nr();
@@ -341,40 +342,40 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2Bui
             self.use_multiple_threads
         );
         let (array, seeds) = if !self.use_multiple_threads {
-            self.best_array(|g| self.conf.build_array_for_hashes(&key_hashes, level_size_segments, level_size_groups as u64, g), level_size_groups)
+            self.best_array(|g| self.conf.build_array_for_hashes(&key_hashes, level_size_segments, level_size_groups, g), level_size_groups)
         } else if level_size_segments >= (1<<17) {
-            self.best_array(|g| self.conf.build_array_for_hashes_mt(&key_hashes, level_size_segments, level_size_groups as u64, g), level_size_groups)
+            self.best_array(|g| self.conf.build_array_for_hashes_mt(&key_hashes, level_size_segments, level_size_groups, g), level_size_groups)
         } else {
-            self.best_array_mt(|g| self.conf.build_array_for_hashes(&key_hashes, level_size_segments, level_size_groups as u64, g), level_size_groups)
+            self.best_array_mt(|g| self.conf.build_array_for_hashes(&key_hashes, level_size_segments, level_size_groups, g), level_size_groups)
         };
         keys.maybe_par_retain_keys_with_indices(
             |i| !array.get_bit(
-                self.conf.hash_index(key_hashes[i], level_size_groups as u64,
+                self.conf.hash_index(key_hashes[i], level_size_groups,
                                      |group| self.conf.bits_per_seed.get_seed(&seeds, group as usize))
             ),
             |key| !array.get_bit(
-                self.conf.key_index(key, level_seed, level_size_groups as u64,
+                self.conf.key_index(key, level_seed, level_size_groups,
                                     |group| self.conf.bits_per_seed.get_seed(&seeds, group as usize))
             ),
             |key| self.retained(key),
             || array.iter().map(|v| v.count_ones() as usize).sum::<usize>(),
             self.use_multiple_threads
         );
-        self.push(array, seeds, level_size_groups as u64);
+        self.push(array, seeds, level_size_groups);
     }
 
-    fn build_next_level<KS, K>(&mut self, keys: &mut KS, level_size_groups: usize, level_size_segments: usize)
+    fn build_next_level<KS, K>(&mut self, keys: &mut KS, level_size_groups: u64, level_size_segments: usize)
         where K: Hash + Sync, KS: KeySet<K> + Sync
     {
         if keys.keys_len() < self.prehash_threshold {
             return self.build_next_level_prehash(keys, level_size_groups, level_size_segments);
         }
         let (array, seeds) = if !self.use_multiple_threads {
-            self.best_array(|g| self.build_array(keys, level_size_segments, level_size_groups as u64, g), level_size_groups)
+            self.best_array(|g| self.build_array(keys, level_size_segments, level_size_groups, g), level_size_groups)
         } else if level_size_segments >= (1<<17) {
-            self.best_array(|g| self.build_array_mt(keys, level_size_segments, level_size_groups as u64, g), level_size_groups)
+            self.best_array(|g| self.build_array_mt(keys, level_size_segments, level_size_groups, g), level_size_groups)
         } else {
-            self.best_array_mt(|g| self.build_array(keys, level_size_segments, level_size_groups as u64, g), level_size_groups)
+            self.best_array_mt(|g| self.build_array(keys, level_size_segments, level_size_groups, g), level_size_groups)
         };
         let level_nr = self.level_nr();
         keys.maybe_par_retain_keys(
@@ -459,7 +460,7 @@ impl<GS: GroupSize + Sync, SS: SeedSize, S: BuildSeededHasher + Sync> FPHash2<GS
                 ceiling_div(input_size * levels.relative_level_size as usize, 100));
             //let seed = level_nr;
             stats.level(input_size, level_size_segments * 64);
-            levels.build_next_level(&mut keys, level_size_groups, level_size_segments);
+            levels.build_next_level(&mut keys, level_size_groups as u64, level_size_segments);
         }
         drop(keys);
         stats.end();
